requirements.txt
addict==2.4.0
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver==2.24.1
click==7.1.2
cryptography==3.4.7
datefinder==0.7.1
Django==3.2.2
idna==2.10
image==1.5.33
mypy-extensions==0.4.3
pathspec==0.8.1
pip-autoremove==0.9.1
pycparser==2.20
python-dateutil==2.8.1
python-magic==0.4.22
pytz==2021.1
razorpay==1.2.0
regex==2021.4.4
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
urllib3==1.26.4
joblib==0.14.1